Convert Identifier.value to short
All possible values fit into an short, so this allows to save
some space. Also re-ordered the fields to avoid alignment holes.

This shrinks the size of an identifier within 16 bytes for -m32 (-m64
remains unchanged at 32 bytes including padding) and hence should consume
less memory for 32bit builds.

+++ src/dmd/frontend.h
@@ -7978,13 +7978,13 @@ struct Id
 
 class Identifier final : public RootObject
 {
-    const int32_t value;
+    const int16_t value;
     const bool isAnonymous_;
     const _d_dynamicArray< const char > name;
 public:
     static Identifier* create(const char* name);
     const char* toChars() const;
-    int32_t getValue() const;
+    int16_t getValue() const;
     bool isAnonymous() const;
     const char* toHChars2() const;
     DYNCAST dyncast() const;
